stabilize Redis caching for APQ and query plan (#2176)
* check for Redis connection issues
* use a null byte as separator for Redis key to prevent potential
injections in the future
* mark all of Redis caching as license key functionality under Elastic
License v2
