+++ apollo-router/src/cache/mod.rs
@@ -9,6 +9,8 @@ use self::storage::CacheStorage;
 use self::storage::KeyType;
 use self::storage::ValueType;
 
+#[cfg(feature = "experimental_cache")]
+mod redis;
 pub(crate) mod storage;
 
 type WaitMap<K, V> = Arc<Mutex<HashMap<K, broadcast::Sender<V>>>>;
@@ -26,24 +28,33 @@ where
     K: KeyType + 'static,
     V: ValueType + 'static,
 {
+    #[cfg(test)]
     pub(crate) async fn new() -> Self {
-        Self::with_capacity(DEFAULT_CACHE_CAPACITY, None).await
+        Self::with_capacity(DEFAULT_CACHE_CAPACITY, None, "test").await
     }
 
-    pub(crate) async fn with_capacity(capacity: usize, redis_urls: Option<Vec<String>>) -> Self {
+    pub(crate) async fn with_capacity(
+        capacity: usize,
+        redis_urls: Option<Vec<String>>,
+        caller: &str,
+    ) -> Self {
         Self {
             wait_map: Arc::new(Mutex::new(HashMap::new())),
-            storage: CacheStorage::new(capacity, redis_urls).await,
+            storage: CacheStorage::new(capacity, redis_urls, caller).await,
         }
     }
 
-    pub(crate) async fn from_configuration(config: &crate::configuration::Cache) -> Self {
+    pub(crate) async fn from_configuration(
+        config: &crate::configuration::Cache,
+        caller: &str,
+    ) -> Self {
         Self::with_capacity(
             config.in_memory.limit,
             #[cfg(feature = "experimental_cache")]
             config.redis.as_ref().map(|c| c.urls.clone()),
             #[cfg(not(feature = "experimental_cache"))]
             None,
+            caller,
         )
         .await
     }
@@ -198,7 +209,7 @@ mod tests {
     #[tokio::test]
     async fn example_cache_usage() {
         let k = "key".to_string();
-        let cache = DeduplicatingCache::with_capacity(1, None).await;
+        let cache = DeduplicatingCache::with_capacity(1, None, "test").await;
 
         let entry = cache.get(&k).await;
 
@@ -215,7 +226,7 @@ mod tests {
     #[test(tokio::test)]
     async fn it_should_enforce_cache_limits() {
         let cache: DeduplicatingCache<usize, usize> =
-            DeduplicatingCache::with_capacity(13, None).await;
+            DeduplicatingCache::with_capacity(13, None, "test").await;
 
         for i in 0..14 {
             let entry = cache.get(&i).await;
@@ -238,7 +249,7 @@ mod tests {
         mock.expect_retrieve().times(1).return_const(1usize);
 
         let cache: DeduplicatingCache<usize, usize> =
-            DeduplicatingCache::with_capacity(10, None).await;
+            DeduplicatingCache::with_capacity(10, None, "test").await;
 
         // Let's trigger 100 concurrent gets of the same value and ensure only
         // one delegated retrieve is made
