+++ apollo-router/src/cache/storage.rs
@@ -1,3 +1,5 @@
+// This entire file is license key functionality
+
 use std::fmt;
 use std::hash::Hash;
 use std::sync::Arc;
@@ -6,6 +8,9 @@ use lru::LruCache;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use tokio::sync::Mutex;
+
+#[cfg(feature = "experimental_cache")]
+use super::redis::*;
 
 pub(crate) trait KeyType:
     Clone + fmt::Debug + fmt::Display + Hash + Eq + Send + Sync
@@ -34,9 +39,6 @@ where
     // It has the functions it needs already
 }
 
-#[cfg(feature = "experimental_cache")]
-use redis_storage::*;
-
 // placeholder storage module
 //
 // this will be replaced by the multi level (in memory + redis/memcached) once we find
@@ -53,12 +55,26 @@ where
     K: KeyType,
     V: ValueType,
 {
-    pub(crate) async fn new(max_capacity: usize, _redis_urls: Option<Vec<String>>) -> Self {
+    pub(crate) async fn new(
+        max_capacity: usize,
+        _redis_urls: Option<Vec<String>>,
+        _caller: &str,
+    ) -> Self {
         Self {
             inner: Arc::new(Mutex::new(LruCache::new(max_capacity))),
             #[cfg(feature = "experimental_cache")]
             redis: if let Some(urls) = _redis_urls {
-                Some(RedisCacheStorage::new(urls).await)
+                match RedisCacheStorage::new(urls).await {
+                    Err(e) => {
+                        tracing::error!(
+                            "could not open connection to Redis for {} caching: {:?}",
+                            _caller,
+                            e
+                        );
+                        None
+                    }
+                    Ok(storage) => Some(storage),
+                }
             } else {
                 None
             },
@@ -101,151 +117,5 @@ where
     #[cfg(test)]
     pub(crate) async fn len(&self) -> usize {
         self.inner.lock().await.len()
-    }
-}
-
-#[cfg(feature = "experimental_cache")]
-mod redis_storage {
-    use std::fmt;
-    use std::sync::Arc;
-
-    use redis::AsyncCommands;
-    use redis::FromRedisValue;
-    use redis::RedisResult;
-    use redis::RedisWrite;
-    use redis::ToRedisArgs;
-    use redis_cluster_async::Client;
-    use redis_cluster_async::Connection;
-    use tokio::sync::Mutex;
-
-    use super::KeyType;
-    use super::ValueType;
-
-    #[derive(Clone, Debug, Eq, Hash, PartialEq)]
-    pub(crate) struct RedisKey<K>(pub(crate) K)
-    where
-        K: KeyType;
-
-    #[derive(Clone, Debug)]
-    pub(crate) struct RedisValue<V>(pub(crate) V)
-    where
-        V: ValueType;
-
-    #[derive(Clone)]
-    pub(crate) struct RedisCacheStorage {
-        inner: Arc<Mutex<Connection>>,
-    }
-
-    fn get_type_of<T>(_: &T) -> &'static str {
-        std::any::type_name::<T>()
-    }
-
-    impl<K> fmt::Display for RedisKey<K>
-    where
-        K: KeyType,
-    {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(f, "{}", self.0)
-        }
-    }
-
-    impl<K> ToRedisArgs for RedisKey<K>
-    where
-        K: KeyType,
-    {
-        fn write_redis_args<W>(&self, out: &mut W)
-        where
-            W: ?Sized + RedisWrite,
-        {
-            out.write_arg_fmt(self);
-        }
-    }
-
-    impl<V> fmt::Display for RedisValue<V>
-    where
-        V: ValueType,
-    {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            write!(f, "{}|{:?}", get_type_of(&self.0), self.0)
-        }
-    }
-
-    impl<V> ToRedisArgs for RedisValue<V>
-    where
-        V: ValueType,
-    {
-        fn write_redis_args<W>(&self, out: &mut W)
-        where
-            W: ?Sized + RedisWrite,
-        {
-            let v = serde_json::to_vec(&self.0).unwrap();
-            out.write_arg(&v);
-        }
-    }
-
-    impl<V> FromRedisValue for RedisValue<V>
-    where
-        V: ValueType,
-    {
-        fn from_redis_value(v: &redis::Value) -> RedisResult<Self> {
-            match v {
-                redis::Value::Bulk(bulk_data) => {
-                    for entry in bulk_data {
-                        tracing::trace!("entry: {:?}", entry);
-                        // entry.parse::<V>().unwrap()
-                    }
-                    Err(redis::RedisError::from((
-                        redis::ErrorKind::TypeError,
-                        "the data is the wrong type",
-                    )))
-                }
-                redis::Value::Data(v) => serde_json::from_slice(v).map(RedisValue).map_err(|e| {
-                    redis::RedisError::from((
-                        redis::ErrorKind::TypeError,
-                        "can't deserialize from JSON",
-                        e.to_string(),
-                    ))
-                }),
-                res => Err(redis::RedisError::from((
-                    redis::ErrorKind::TypeError,
-                    "the data is the wrong type",
-                    format!("{:?}", res),
-                ))),
-            }
-        }
-    }
-
-    impl RedisCacheStorage {
-        pub(crate) async fn new(urls: Vec<String>) -> Self {
-            let client = Client::open(urls).expect("opening ClusterClient");
-            let connection = client.get_connection().await.expect("got redis connection");
-
-            tracing::trace!("redis connection established");
-            Self {
-                inner: Arc::new(Mutex::new(connection)),
-            }
-        }
-
-        pub(crate) async fn get<K: KeyType, V: ValueType>(
-            &self,
-            key: RedisKey<K>,
-        ) -> Option<RedisValue<V>> {
-            tracing::trace!("getting from redis: {:?}", key);
-            let mut guard = self.inner.lock().await;
-            guard.get(key).await.ok()
-        }
-
-        pub(crate) async fn insert<K: KeyType, V: ValueType>(
-            &self,
-            key: RedisKey<K>,
-            value: RedisValue<V>,
-        ) {
-            tracing::trace!("inserting into redis: {:?}, {:?}", key, value);
-            let mut guard = self.inner.lock().await;
-            let r = guard
-                .set::<RedisKey<K>, RedisValue<V>, redis::Value>(key, value)
-                .await;
-            tracing::trace!("insert result {:?}", r);
-        }
     }
 }
